+++ tests/master.lisp
@@ -17,14 +17,14 @@
 (defun fn3 (item)
   (format nil "test3 ~a" item))
 
-(define-system
+(define-system ()
     (:name test-node1 :fn #'fn1 :batch-size 5)
     (:name test-node2 :fn #'fn2 :batch-size 10)
     (:name test-node3 :fn #'fn3 :batch-size 4))
 
-(define-rmq-node final-work-node nil queue-4 queue-4 10)
+(define-rmq-node final-work-node nil queue-4 10 :dest-queue queue-4)
 
-(define-rmq-node final-work-node1 nil queue-3 queue-3 10)
+(define-rmq-node final-work-node1 nil queue-3 10 :dest-queue queue-3)
 
 (teardown
   (let ((conn (setup-connection :host *rmq-host* :username *rmq-user*